pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "asr4"
dynamic = ["readme", "description", "version"]
requires-python = ">=3.7,!=3.11"
keywords = ["ASR", "CSR"]
license = {text = "BSD 3-Clause License"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Telecommunications Industry",
    "Natural Language :: English",
    "Natural Language :: Spanish",
    "Natural Language :: Portuguese (Brazilian)",
    "License :: Other/Proprietary License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
]
dependencies = [
    "google > =3.0.0",
    "grpcio >= 1.47.0",
    "protobuf >= 3.20.1",
    "typing_extensions >= 4.3.0",
    "logger >= 1.4",
]

[tool.setuptools]
include-package-data = false

[tool.setuptools.package-data]
asr4 = ["../DESCRIPTION.md", "../VERSION"]

[tool.setuptools.packages.find]
exclude = ["tests", "tests.*"]

[tool.setuptools.dynamic]
readme = {file = ["README.md"]}
description = {file = ["DESCRIPTION.md"]}
version = {file = ["VERSION"]}

[project.optional-dependencies]
client = [
    "csr-evaluation-tools @ https://asr4:glpat-VkJr-mymfHB2P8GrLzvb@gitlab.lan.verbio.com/api/v4/projects/1420/packages/pypi/files/b82a19177ce0fb48f241ce9a026be6012958d4449593f7fec850df83bc0e3577/csr_evaluation_tools-0.0.3-py37-none-linux_x86_64.whl ; python_version=='3.7' and platform_system=='Linux'",
    "csr-evaluation-tools @ https://asr4:glpat-VkJr-mymfHB2P8GrLzvb@gitlab.lan.verbio.com/api/v4/projects/1420/packages/pypi/files/937ab0f1db46876e937cb89adbf2233eaca5cd6f0696fb7d01b7e461e10fd210/csr_evaluation_tools-0.0.3-py38-none-linux_x86_64.whl ; python_version=='3.8' and platform_system=='Linux'",
    "csr-evaluation-tools @ https://asr4:glpat-VkJr-mymfHB2P8GrLzvb@gitlab.lan.verbio.com/api/v4/projects/1420/packages/pypi/files/3b0b5cb71b6dbe9288ba2c1418bea0947cb1616a726a2c3c1a0debda9798ae9f/csr_evaluation_tools-0.0.3-py39-none-linux_x86_64.whl ; python_version=='3.9' and platform_system=='Linux'",
    "csr-evaluation-tools @ https://asr4:glpat-VkJr-mymfHB2P8GrLzvb@gitlab.lan.verbio.com/api/v4/projects/1420/packages/pypi/files/b32f3244fba6ecbf47102214a42c2f12c54cd55016e8c0d636c3329460f5ca04/csr_evaluation_tools-0.0.3-py310-none-linux_x86_64.whl ; python_version=='3.10' and platform_system=='Linux'",
]
server = [
    "resampy >= 0.4.2",
    "grpcio-health-checking >= 1.47.0",
    "pyformatter @ https://__token__:qKtzEZznR88brmJub-nm@gitlab.lan.verbio.com/api/v4/projects/949/packages/pypi/files/dd9cfb194c7c3c4652e1a320dfadf71d00db75c3081aa554f3b31bc4faf9f4b0/pyformatter-1.3.11-cp311-cp311-linux_x86_64.whl ; python_version=='3.11' and platform_system=='Linux'",
    "pyformatter @ https://__token__:qKtzEZznR88brmJub-nm@gitlab.lan.verbio.com/api/v4/projects/949/packages/pypi/files/789824799d8386891c8dfefd81968d714eaa804eb5a3c43d32b996c37a091f99/pyformatter-1.3.11-cp310-cp310-linux_x86_64.whl ; python_version=='3.10' and platform_system=='Linux'",
    "pyformatter @ https://__token__:qKtzEZznR88brmJub-nm@gitlab.lan.verbio.com/api/v4/projects/949/packages/pypi/files/171256e80601d9542833c68b4245cd6b089c5eeabdde630f3385b93d4a8909aa/pyformatter-1.3.11-cp39-cp39-linux_x86_64.whl ; python_version=='3.9' and platform_system=='Linux'",
    "pyformatter @ https://__token__:qKtzEZznR88brmJub-nm@gitlab.lan.verbio.com/api/v4/projects/949/packages/pypi/files/c2efb5f5335d12f986c6eac5dd6532cc41779b0b544e2705f24b9c2763d20014/pyformatter-1.3.11-cp38-cp38-linux_x86_64.whl ; python_version=='3.8' and platform_system=='Linux'",
    "pyformatter @ https://__token__:qKtzEZznR88brmJub-nm@gitlab.lan.verbio.com/api/v4/projects/949/packages/pypi/files/9d7278a1d271c73e1879a8d9e0d1981fbc998625bb3bee68584c0f620a537012/pyformatter-1.3.11-cp37-cp37m-linux_x86_64.whl ; python_version=='3.7' and platform_system=='Linux'",
    "pyformatter @ https://__token__:qKtzEZznR88brmJub-nm@gitlab.lan.verbio.com/api/v4/projects/949/packages/pypi/files/7c7b0b4600cf3d5b40d2383ea7c1af31a65940c2cd55df9e88517f5f27bc6ce3/pyformatter-1.3.11-cp36-cp36m-linux_x86_64.whl ; python_version=='3.6' and platform_system=='Linux'",
    "simple-ctc @ git+https://github.com/mthrok/ctcdecode@b1a30d7a65342012e0d2524d9bae1c5412b24a23"
]
cpu = [
    "torch @ https://download.pytorch.org/whl/cpu/torch-1.12.1%2Bcpu-cp37-cp37m-linux_x86_64.whl ; python_version=='3.7' and platform_system=='Linux'",
    "torch @ https://download.pytorch.org/whl/cpu/torch-1.12.1%2Bcpu-cp38-cp38-linux_x86_64.whl ; python_version=='3.8' and platform_system=='Linux'",
    "torch @ https://download.pytorch.org/whl/cpu/torch-1.12.1%2Bcpu-cp39-cp39-linux_x86_64.whl ; python_version=='3.9' and platform_system=='Linux'",
    "torch @ https://download.pytorch.org/whl/cpu/torch-1.12.1%2Bcpu-cp310-cp310-linux_x86_64.whl ; python_version=='3.10' and platform_system=='Linux'",
    "torch @ https://download.pytorch.org/whl/cpu/torch-1.12.1%2Bcpu-cp37-cp37m-win_amd64.whl ; python_version=='3.7' and platform_system=='Windows'",
    "torch @ https://download.pytorch.org/whl/cpu/torch-1.12.1%2Bcpu-cp38-cp38-win_amd64.whl ; python_version=='3.8' and platform_system=='Windows'",
    "torch @ https://download.pytorch.org/whl/cpu/torch-1.12.1%2Bcpu-cp39-cp39-win_amd64.whl ; python_version=='3.9' and platform_system=='Windows'",
    "torch @ https://download.pytorch.org/whl/cpu/torch-1.12.1%2Bcpu-cp310-cp310-win_amd64.whl ; python_version=='3.10' and platform_system=='Windows'",
    "onnxruntime >= 1.12.1",
]
gpu = [
    "torch == 1.12.0a0+git8c5c07c",
    "onnxruntime-gpu >= 1.12.1",
]
test = [
    "pytest >= 6.2.5",
    "pytest-cov >= 3.0.0",
    "jiwer >= 2.5.1"
]